+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Shape#clear_shapes just like that of Drawable in WidgetProxy
 - Speed up glimmer command startup time by switching to a shell script
 - glimmer-setup command installs `glimmer` and `girb` as shell commands
+- Fix issue with processing arguments for `glimmer package` command.
 
 ### 4.18.7.7